move auth code to  package
